core/pyproject.toml
[tool.poetry]
name = "poetry-project"
version = "0.1.0"
description = ""
authors = ["iMaksus <max1112vas@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry_project"}]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
python-dotenv = "^0.21.0"
djangorestframework-jwt = "^1.11.0"
djoser = "^2.1.0"
djangorestframework-simplejwt = "~4.6.0"
PyJWT = "^1.5.2"
drf-nested-routers = "^0.93.4"
boto3 = "^1.25.4"
botocore = "^1.28.4"


[build-system]
# requires = ["poetry-core"]
requires = ["poetry>=1.2"]
build-backend = "poetry.core.masonry.api"
